+++ CHANGELOG.md
@@ -873,6 +873,7 @@
 
 - Record changes to refunds and show them in the activity's Change history
 - Show the commitment value on an activity's finance tab
+- Include any commitment in an activity's IATI XML
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-80...HEAD
 [release-80]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-79...release-80